Don't compare result of compareTo() against values other than 0.
BigInteger.compareTo() happens to guarantee that it only returns -1, 0, and 1, but that's not the case for compareTo() in general. This makes the old code harder to read / fragile. This commit replaces comparisons == -1 with an equivalent comparison < 0.

I haven't tried to compile or run any tests, since I don't have a dev environment for this set up.