Remove the UNALLOCAED special type
Get rid of the enum variant used for unallocated special types as it
cannot be decoded or encoded.

Move the decoder exception to be thrown when trying to parse an
unallocated special type rather than returning UNALLOCAED and then
throwing the exception.

Remove the handling of this enum varient in the encoder which is also
made to throw an AssertionError in the accidental case of a
non-exhaustive switch statement.