Remove the INVALID major type
Get rid of the enum variant that is used for invalid major types that
cannot be decoded or encoded.

Move the decoder exception to be thrown when trying to parse the invalid
major type rather than returning INVALID and then throwing the
exception.

Remove handling of this enum variant in the encoder which is also make
to throw an AssertionError in the accidental case of a
non-exhaustive switch statement.