Introduce helper method to encode to bytes
It's common to encode CBOR to a byte array so provide utility methods in
CborEncoder and on DataItem to make it really simple.

These methods don't throw exceptions as only exceptions from the encoder
will be from the OutputStream but a ByteArrayOutputStream doesn't
actually throw so the exception can be ignored. An AssertionError is
thrown in the unexpected event that an exception did somehow get caught.

The alternative to ignoring the exception would be to have a shadow
implementation of the encoder that directly uses the
ByteArrayOutputStream methods that are declared not to throw so the
shadow implementatino can also declare methods that don't throw. This
would result in duplication so it didn't seem worth the maintenance
cost.